Use ES6 collections for lookups.
In a very large JSON schema object (>7MiB), it was found that the
overwhelming majority of time was spent on O(N) lookups in the
context array. Adding a Set in parallel vastly improves performance,
even though the array is kept as a duplicate in order to derive the
proper number for hashing cyclic objects.